Update patch set 22
Patch Set 22: Code-Review-1

(7 comments)

> Is there something special we need to do to get a dep from fedora
 > available to RDO?
 > 
 > I keep hitting this: Error: No Package found for python2-tosca-parser
 > 
 > In fedora I can see:
 > dnf search tosca:
 > python2-tosca-parser.noarch : Parser for TOSCA Simple Profile in
 > YAML

This is an OpenStack project, so we can package it in DLRN or include it in the dependency repo (this has been done a few times, e.g. for the XStatic packages).

If this is only a test requirement, and not an optional requiement we will need to use (because we import TOSCA templates for some reason), we might prefer to skip it and not run tests until it's packaged.

Patch-set: 22
Label: Code-Review=-1