Empty changelog, force a package rebuild
We've had an issue in the centos-master worker, because we first
had heat-translator unpinned, but then it was pinned to version 0.6.0.
After the 0.7.0 version was pinned, it would not be rebuilt because
it had already been before, so the older package (0.6.0) stayed.

We need to rebuild, so we can take the chance to empty the changelog,
according to the DLRN requirements.

Change-Id: Ic0a154063f8c4e3fffd6b2848cf21088d25e9cf5